pyproject.toml
[tool.poetry]
name = "unterricht-python-backend"
version = "0.1.0"
description = ""
authors = ["CTS <cts_anubis@gmx.de>"]
readme = "README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.0"
langchain-community = "^0.0.16"
langchain-openai = "*"
Flask = "^3.0.0"
ConfigParser = "*"
pypdf = "*"
openai = "*"
tiktoken = "*"
python-dotenv = "*"
faiss-cpu = "*"
supabase = "*"
werkzeug = "*"
scikit-learn = "*"
gunicorn = "*"
pydantic = "*"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
